Add meshutils.js and shift into it part of gfxutils
